handled Exceptions and Returning appropeiate HttpStatus code

+++ src/main/java/org/scaler/productservice/service/localDbServices/SelfProductService.java
@@ -1,5 +1,6 @@
 package org.scaler.productservice.service.localDbServices;
 
+import org.scaler.productservice.exceptions.NotFoundException;
 import org.scaler.productservice.models.Category;
 import org.scaler.productservice.models.Product;
 import org.scaler.productservice.repository.CategoryRepository;
@@ -8,6 +9,7 @@ import org.scaler.productservice.service.Productservice;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 
 @Service("selfProductService")
@@ -22,13 +24,21 @@ public class SelfProductService implements Productservice {
     }
 
     @Override
-    public Product getProductById(Long id) {
-        return productRepository.findByIdIs(id);
+    public Product getProductById(Long id) throws NotFoundException {
+        Optional<Product> response = productRepository.findById(id);
+        if(response.isPresent()){
+            return response.get();
+        }
+        throw new NotFoundException("Product with id:"+id+" doesn't exist");
     }
 
     @Override
-    public List<Product> getAllProduct() {
-        return productRepository.findAll();
+    public List<Product> getAllProduct() throws NotFoundException {
+        List<Product> response = productRepository.findAll();
+        if(!response.isEmpty()){
+            return response;
+        }
+        throw new NotFoundException("There is no products");
     }
 
     @Override
@@ -49,45 +59,54 @@ public class SelfProductService implements Productservice {
     }
 
     @Override
-    public Product deleteProduct(Long id) {
-        Product productFromDb = productRepository.findByIdIs(id);
-        productRepository.deleteById(id);
-        return productFromDb;
+    public Product deleteProduct(Long id) throws NotFoundException {
+        Optional<Product> productFromDb = productRepository.findById(id);
+        if(productFromDb.isPresent()) {
+            productRepository.deleteById(id);
+            return productFromDb.get();
+        }
+        throw new NotFoundException("Product with id:"+id+" doesn't exist");
     }
 
     @Override
-    public Product updateProduct(Long id, String title, Double price, String description, String image, String categoryName) {
-        Product currentProduct = productRepository.findByIdIs(id);
-        if(title != null){
-            currentProduct.setTitle(title);
-        }
-        if(price != null){
-            currentProduct.setPrice(price);
-        }
-        if(description != null){
-            currentProduct.setDescription(description);
-        }
-        if(image != null){
-            currentProduct.setImage(image);
-        }
-        if(categoryName == null){
-            currentProduct.setCategory(currentProduct.getCategory());
-        }
-        else {
-            Category categoryFromDb = categoryRepository.findByTitle(categoryName);
-            if (categoryFromDb == null) {
-                categoryFromDb = new Category();
-                categoryFromDb.setTitle(categoryName);
+    public Product updateProduct(Long id, String title, Double price, String description, String image, String categoryName) throws NotFoundException {
+        Optional<Product> response = productRepository.findById(id);
+        if(response.isPresent()) {
+            Product currentProduct = response.get();
+            if (title != null) {
+                currentProduct.setTitle(title);
             }
-            currentProduct.setCategory(categoryFromDb);
+            if (price != null) {
+                currentProduct.setPrice(price);
+            }
+            if (description != null) {
+                currentProduct.setDescription(description);
+            }
+            if (image != null) {
+                currentProduct.setImage(image);
+            }
+            if (categoryName == null) {
+                currentProduct.setCategory(currentProduct.getCategory());
+            } else {
+                Category categoryFromDb = categoryRepository.findByTitle(categoryName);
+                if (categoryFromDb == null) {
+                    categoryFromDb = new Category();
+                    categoryFromDb.setTitle(categoryName);
+                }
+                currentProduct.setCategory(categoryFromDb);
+            }
+            return productRepository.save(currentProduct);
         }
 
-
-        return productRepository.save(currentProduct);
+        throw new NotFoundException("Product with id:"+id+" doesn't exist");
     }
 
     @Override
-    public List<Product> getAllProductByCategory(String title) {
-        return productRepository.findByCategoryTitle(title);
+    public List<Product> getAllProductByCategory(String title) throws NotFoundException {
+        List<Product> response = productRepository.findByCategoryTitle(title);
+        if(!response.isEmpty()){
+            return response;
+        }
+        throw new NotFoundException("There is no product in "+title+" category");
     }
 }
